+++ packages/plugin-text-color/src/createTextColorPlugin.js
@@ -1,15 +1,7 @@
 import { createBasePlugin } from 'wix-rich-content-common';
 import { TEXT_COLOR_TYPE } from './types';
-import { isHexColor } from './utils';
 import createTextColorToolbar from './toolbar/createTextColorToolbar';
-
-const customStyleFn = styles =>
-  styles
-    .toArray()
-    .reduce(
-      (cssStyle, style) => ({ ...cssStyle, ...(isHexColor(style) ? { color: style } : {}) }),
-      {}
-    );
+import { DEFAULT_STYLE_FN_DRAFT } from './constants';
 
 const createTextColorPlugin = (config = {}) => {
   const type = TEXT_COLOR_TYPE;
@@ -21,7 +13,7 @@ const createTextColorPlugin = (config = {}) => {
     toolbar,
     type,
     settings,
-    customStyleFn,
+    customStyleFn: settings.customStyleFn || DEFAULT_STYLE_FN_DRAFT,
     ...rest,
   });
 };