+++ packages/plugin-text-color/src/toolbar/TextColorButton.jsx
@@ -3,9 +3,8 @@ import PropTypes from 'prop-types';
 import { getModalStyles, InlineToolbarButton, getSelectionStyles } from 'wix-rich-content-common';
 import TextColorIcon from './TextColorIcon';
 import { TEXT_COLOR_TYPE } from '../types';
-import { MODAL_STYLES, PANEL_WIDTH } from './constants';
+import { MODAL_STYLES, PANEL_WIDTH, DEFAULT_STYLE_SELECTION_PREDICATE } from '../constants';
 import { Modals } from '../modals';
-import { isHexColor } from '../utils';
 
 export default class TextColorButton extends Component {
   constructor(props) {
@@ -81,7 +80,10 @@ export default class TextColorButton extends Component {
   };
 
   get isActive() {
-    return getSelectionStyles(style => isHexColor(style), this.props.getEditorState()).length > 0;
+    const settings = this.props.config[TEXT_COLOR_TYPE] || {};
+    const styleSelectionPredicate =
+      settings.styleSelectionPredicate || DEFAULT_STYLE_SELECTION_PREDICATE;
+    return getSelectionStyles(styleSelectionPredicate, this.props.getEditorState()).length > 0;
   }
 
   render() {